Move types to separate files and rename to match type name
