+++ src/System.Text.JsonLab/System/Text/Json/JsonParser.cs
@@ -2,137 +2,10 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Buffers;
-using System.Diagnostics;
 using System.Runtime.InteropServices;
 
 namespace System.Text.JsonLab
 {
-    // Location - offset - 0 - size - 4
-    // Length - offset - 4 - size - 4
-    // Type - offset - 8 - size - 1
-    [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    internal struct DbRow
-    {
-        public static int Size;
-
-        public int Location;                // index in JSON payload
-        public int Length;      // length of text in JSON payload
-        public JsonValueType Type; // type of JSON construct (e.g. Object, Array, Number)
-
-        public const int UnknownNumberOfRows = -1;
-
-        public DbRow(JsonValueType type, int valueIndex, int lengthOrNumberOfRows = UnknownNumberOfRows)
-        {
-            Location = valueIndex;
-            Length = lengthOrNumberOfRows;
-            Type = type;
-        }
-
-        public bool IsSimpleValue => Type > JsonValueType.Array;
-
-        unsafe static DbRow()
-        {
-            Size = sizeof(DbRow);
-        }
-    }
-
-    // IsArray - offset - 0 - size - 1
-    // Length - offset - 1 - size - 4
-    [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    internal struct StackRow
-    {
-        public static int Size;
-
-        public bool IsArray;
-        public int Length;
-
-        public StackRow(bool isArray, int lengthOrNumberOfRows)
-        {
-            IsArray = isArray;
-            Length = lengthOrNumberOfRows;
-        }
-
-        unsafe static StackRow()
-        {
-            Size = sizeof(StackRow);
-        }
-    }
-
-    internal ref struct CustomStack
-    {
-        private Span<byte> _stackSpace;
-        private int _topOfStack;
-
-        public CustomStack(Span<byte> stackSpace)
-        {
-            _stackSpace = stackSpace;
-            _topOfStack = stackSpace.Length;
-        }
-
-        public bool TryPush(StackRow row)
-        {
-            if (_topOfStack >= StackRow.Size)
-            {
-                MemoryMarshal.Write(_stackSpace.Slice(_topOfStack - StackRow.Size), ref row);
-                _topOfStack -= StackRow.Size;
-                return true;
-            }
-            return false;
-        }
-
-        public StackRow Pop()
-        {
-            StackRow row = Peek();
-            _topOfStack += StackRow.Size;
-            return row;
-        }
-
-        public StackRow Peek()
-        {
-            if (_topOfStack > _stackSpace.Length - StackRow.Size)
-            {
-                JsonThrowHelper.ThrowInvalidOperationException();
-            }
-
-            return MemoryMarshal.Read<StackRow>(_stackSpace.Slice(_topOfStack));
-        }
-
-        public bool IsTopArray()
-        {
-            if (_topOfStack > _stackSpace.Length - StackRow.Size)
-            {
-                return false;
-            }
-            return _stackSpace[_topOfStack] == 1;
-        }
-
-        private void Reset()
-        {
-            _stackSpace.Slice(_topOfStack - StackRow.Size, StackRow.Size).Fill(255);
-        }
-
-        public void Resize(Span<byte> newStackMemory)
-        {
-            Debug.Assert(newStackMemory.Length > _stackSpace.Length);
-            int stackGrowth = newStackMemory.Length - _stackSpace.Length;
-            _stackSpace.CopyTo(newStackMemory.Slice(stackGrowth));
-            _topOfStack += stackGrowth;
-            _stackSpace = newStackMemory;
-        }
-
-        public string PrintStacks()
-        {
-            StringBuilder sb = new StringBuilder();
-            sb.Append("IsArray" + "\t" + "Length" + "\r\n");
-            for (int i = _stackSpace.Length - StackRow.Size; i >= StackRow.Size; i -= StackRow.Size)
-            {
-                StackRow row = MemoryMarshal.Read<StackRow>(_stackSpace.Slice(i));
-                sb.Append(row.IsArray + "\t" + row.Length + "\r\n");
-            }
-            return sb.ToString();
-        }
-    }
-
     public ref struct JsonParser
     {
         private readonly ReadOnlySpan<byte> _utf8Json; // TODO: this should be ReadOnlyMemory<byte>
